gradle-plugin: fix docs and tests for Groovy DSL
### What's done:
* Update build.gradle in examples
* Add test for Groovy DSL with explicit inputs
* Minor cleanup in gradle-plugin's docs
